// src/terrain/types.ts
export type DTEDLevelName = 'dt0' | 'dt1' | 'dt2';
export type DTEDMode = 'local' | 'remote' | 'hybrid';

export interface DTEDHeader {
  originLon: number;
  originLat: number;
  lonInterval: number;
  latInterval: number;
  numLonLines: number;
  numLatPoints: number;
}

export interface DTEDTile {
  id: string;
  level: DTEDLevelName;
  origin: [number, number];
  width: number;
  height: number;
  elevations: Int16Array;
  minElevation: number;
  maxElevation: number;
  extent: {
    minX: number;
    minY: number;
    maxX: number;
    maxY: number;
  };
}